refa(portal): 将 repoId 改为 string 类型
Local vcs 只有 path 来区分无 id，所以将 repoId 改为了字符串，
根据不同的 vcs 类型可能是一个数值 id 或者是一个表示 path 的字符串

涉及表结构字段改动，表单字段类型改动，接口返回值字段类型改动

+++ apps/open_template.go
@@ -73,7 +73,7 @@ func OpenDetailTemplate(c *ctx.ServiceCtx, gUid string) (interface{}, e.Error) {
 	}
 
 	if vcs.VcsType == consts.GitTypeGitEA {
-		commit,err:= vcs2.GetGiteaBranchCommitId(vcs,uint(tpl.RepoId),tpl.RepoBranch)
+		commit, err := vcs2.GetGiteaBranchCommitId(vcs, tpl.RepoId, tpl.RepoBranch)
 		if err != nil {
 			return nil, e.New(e.GitLabError, fmt.Errorf("query commit id error: %v", er))
 		}
